+++ extras/tests/Numbers/parseInteger.cpp
@@ -3,8 +3,7 @@
 // MIT License
 
 #include <stdint.h>
-#include <ArduinoJson/Numbers/parseNumber.hpp>
-#include <ArduinoJson/Variant/VariantImpl.hpp>
+#include <ArduinoJson.hpp>
 #include <catch.hpp>
 
 using namespace ARDUINOJSON_NAMESPACE;