+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ArduinoJson: change log
 HEAD
 ----
 
+* Added support for custom converters (issue #687)
 * Removed support for `char` values, see below (issue #1498)
 * `deserializeJson()` leaves `\uXXXX` unchanged instead of returning `NotSupported`
 * `deserializeMsgPack()` inserts `null` instead of returning `NotSupported`